Move NodeDataValidator to designer module
