+++ scenario-compiler/src/main/scala/pl/touk/nussknacker/engine/compile/nodecompilation/NodeCompiler.scala
@@ -23,7 +23,6 @@ import pl.touk.nussknacker.engine.compile.nodecompilation.NodeCompiler.{
   MockExpressionParameterName,
   NodeCompilationResult
 }
-import pl.touk.nussknacker.engine.compile.nodecompilation.NodeDataValidator.OutgoingEdge
 import pl.touk.nussknacker.engine.compiledgraph.{CompiledParameter, TypedParameter}
 import pl.touk.nussknacker.engine.compiledgraph.service.ServiceRef
 import pl.touk.nussknacker.engine.compiledgraph.variable.Field
@@ -489,7 +488,7 @@ class NodeCompiler(
     }
   }
 
-  private[compile] def compileFragmentInput(
+  def compileFragmentInput(
       fragmentInput: FragmentInput,
       inputContext: SingleInputNodeInputValidationContext
   )(
