+++ scenario-compiler/src/main/scala/pl/touk/nussknacker/engine/compile/nodecompilation/package.scala
@@ -1,0 +1,7 @@
+package pl.touk.nussknacker.engine.compile
+
+import pl.touk.nussknacker.engine.graph.EdgeType
+
+package object nodecompilation {
+  case class OutgoingEdge(target: String, edgeType: Option[EdgeType])
+}
